fix(x64): repair VC6 build broken by BaseTypeCore.h, fix G3's incremental-build blind spot
G3, the gate proving 64-bit porting work leaves the VC6 retail build untouched,
reported "7 of 7 .text digests match" at nine consecutive tasks. All nine passes
were false. The baseline it compared against was recorded from a --clean build;
every G3 re-run since used an incremental build. Files nobody touched kept their
previously-built .obj, and a previously-built .obj cannot produce a different
digest no matter what happened elsewhere in the tree — so the gate was comparing
yesterday's object code to itself and calling the match proof.

Two real problems hid behind those false passes:

1. The VC6 build had been broken since a task added
   #include "Lib/BaseTypeCore.h" to debug_debug.h to reach UnsignedIntPtr.
   debug_debug.h is transitively included by nearly every debug translation
   unit, including debug_io_flat.cpp, whose `unsigned pathLen=strlen(path);`
   (line 80, unused, harmless under VC6 since before this branch) inherited
   BaseTypeCore.h's `#pragma warning(error : 4189)` (BaseTypeCore.h:80; also
   4706 at :67 and 4101 at :83) and became a hard compile error:

     debug_io_flat.cpp(80): error C4189: 'pathLen' : local variable is
     initialized but not referenced

   GCC ignores #pragma warning entirely, so neither MinGW gate (i686 or x64)
   could ever see this — only VC6 could, and VC6 is exactly the gate the
   incremental-build blind spot above had defanged. Confirmed by clean
   --clean builds of both games from the pre-fix tree, both failing with
   this identical error.

2. VC6 codegen had genuinely shifted during the fix wave that repairs (1),
   and an incremental gate structurally cannot detect that either: it never
   recompiles the files whose codegen moved unless something else already
   forced a rebuild.

Fix wave (the 20 files below):
- debug_debug.{h,cpp}, debug_stack.{h,cpp}, WWLib/Except.{h,cpp},
  WWLib/registry.{h,cpp}, WWSaveLoad/persistfactory.h, WWAudio/{AudibleSound.
  {h,cpp},SoundScene.cpp,SoundSceneObj.h}, profile/profile_funclevel.h:
  swapped `Lib/BaseTypeCore.h` for `<Utility/stdint_adapter.h>` +
  `uintptr_t`, removing the warning-as-error pragmas from files that never
  had them (this branch's own established precedent, already followed in
  surfaceclass.cpp/dx8webbrowser.cpp; now the documented reason to keep
  following it).
- Except.cpp's stack dump and debug_stack.cpp's Signature addressing read/
  printed x64 stack slots and addresses in 4-byte units (truncating crash
  reports); widened to uintptr_t with width-correct format specifiers.
  debug_debug.cpp's pointer/address printing had the same 32-bit-only
  truncation; same fix. debug_debug.cpp's SkipNext() 32-bit inline-asm arm
  was restored to match its siblings' codegen exactly (a prior task had
  collapsed it into an ungated builtin, silently changing 32-bit GCC/Clang
  codegen).
- rendobj.cpp, both dazzle.cpp copies, and AudibleSound.cpp's Load paths
  read pointer-sized (8-byte) tokens from legacy 4-byte W3D micro-chunks on
  x86-64, silently poisoning SaveLoadSystemClass's pointer-remap table
  (ChunkLoadClass::Read refuses oversized reads and leaves the destination
  untouched, rather than failing loudly). This is a real, live defect: the
  read side has an actual in-repo caller (W3DView's
  AnimatedSoundOptionsDialog.cpp), documented in this pass's addendum to
  savegame-format-decision.md. Fixed to read a fixed 4-byte token and cast
  after, mirroring persistfactory.h's existing pattern. The matching Save
  halves remain unreachable and are documented, not fixed, per that same
  addendum.
- cmake/dx8.cmake: gated d3dx8's link (pe-i386-only, no x64 build exists) on
  CMAKE_SIZEOF_VOID_P EQUAL 4, matching its sibling gate.
- docs/x64/core-error-catalogue-v2.md: marked v1's unreproducible
  266-translation-unit coverage figure provenance-uncertain rather than
  silently repeating or discarding it.

VC6 codegen intentionally moved for the 5 of 7 Zero Hour executables (and
their Generals counterparts) that link this changed code, and that change is
accepted, not reverted: every fix above repairs a real defect that also
existed on 32-bit, just less visibly (a 4-byte-truncated address is still
correct where addresses are 4 bytes). These are not architecture-conditional
bugs, so their fixes aren't either. Preserving byte-identical .text would
mean knowingly shipping broken crash dumps to keep a hash stable. Upstream's
constraint (issue #473) is "must not break the VS6 build and compatibility"
— the build works, and the change is confined to crash reporting, audio
handle plumbing, registry access, W3D-asset load paths, profiling, and one
CMake link gate. None of it touches game logic, the actual gameplay
savegame system (Common/Xfer.h and friends — untouched; WWSaveLoad here is
a separate, older W3D-asset framework), or network code — verified against
the file list, not assumed. imagepacker.exe and wdump.exe, which link none
of the changed code, reproduce the pre-fix-wave .text/normalized digests
byte-for-byte, independently confirming the measurement pipeline.

Clean-build gate results at this commit:
- VC6 Zero Hour, --clean: exit 0, 0 errors
- VC6 Generals, --clean: exit 0, 0 errors
- MinGW i686 control, clean full: exit 0, 0 errors, 31,555 warnings —
  matches docs/x64/baseline-mingw-i686.md
- MinGW x64, clean full: 602 errors / 36 shapes / 102 files / 1,669
  objects — unchanged, as expected (correctness fixes, not error-count
  fixes)

docs/x64/baseline-vc6.md rewritten: the incremental-vs-clean trap moved to
the top, the C4189 chain and BaseTypeCore.h precedent added, both games'
.text tables recorded (Zero Hour's baseline previously existed; Generals
never had one despite Core changes since needing it), existing
normalized-hash/git-metadata-drift mechanics preserved.

scripts/verify-retail-baseline.sh added: always clean-builds both games and
diffs .text against the doc (no incremental code path exists to forget to
disable), with --check-only to compare already-built artifacts without
triggering a build. Verified by parsing/hashing check against the on-disk
clean-build artifacts (--check-only, no build run) and a negative-path test
against mismatched artifacts to confirm mismatch/missing-file detection and
non-zero exit both work.

+++ Core/Libraries/Source/WWVegas/WWLib/Except.h
@@ -39,6 +39,7 @@
 #if defined(_WIN32)
 
 #include "win.h"
+#include <Utility/stdint_adapter.h>
 /*
 ** Forward Declarations
 */
@@ -46,7 +47,12 @@ typedef struct _EXCEPTION_POINTERS EXCEPTION_POINTERS;
 typedef struct _CONTEXT CONTEXT;
 
 int Exception_Handler(int exception_code, EXCEPTION_POINTERS *e_info);
-int Stack_Walk(unsigned long *return_addresses, int num_addresses, CONTEXT *context = nullptr);
+// return_addresses is uintptr_t, not unsigned long: it holds raw return
+// addresses captured off the stack (see Except.cpp's Stack_Walk), and on
+// x86-64 (LLP64) `long` stays 32 bits while an address is 64. Stack_Walk has
+// exactly one call site (Except.cpp), so this is a self-contained widening,
+// not a public-ABI change in practice.
+int Stack_Walk(uintptr_t *return_addresses, int num_addresses, CONTEXT *context = nullptr);
 bool Lookup_Symbol(void *code_ptr, char *symbol, int &displacement);
 void Load_Image_Helper();
 void Register_Thread_ID(unsigned long thread_id, char *thread_name, bool main = false);
